tpm2: split version-specific code out of policy.go
This doesn't update any tests yet.
