+++ tpm2/keydata_v1_test.go
@@ -81,7 +81,7 @@ func (s *keyDataV1Suite) newMockKeyData(c *C, pcrPolicyCounterHandle tpm2.Handle
 	template := tpm2_testutil.NewSealedObjectTemplate()
 
 	trial := util.ComputeAuthPolicy(template.NameAlg)
-	trial.PolicyAuthorize(ComputePcrPolicyRefFromCounterName(policyCounterName), authKeyName)
+	trial.PolicyAuthorize(ComputeV1PcrPolicyRefFromCounterName(policyCounterName), authKeyName)
 	trial.PolicyAuthValue()
 
 	template.AuthPolicy = trial.GetDigest()
